requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
h11==0.16.0
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
sqlalchemy==1.4.49
psycopg2-binary==2.9.6
pydantic[email]==2.11.7
PyJWT>=2.0.0
google-api-python-client
google-auth
google-auth-httplib2
openpyxl
httpx==0.28.1
python-dotenv==1.0.1
APScheduler==3.11.0
reportlab==4.2.5
